+++ packages/app-headless-cms/src/admin/components/ContentModelEditor/Tabs/PreviewTab/PreviewTab.tsx
@@ -5,6 +5,8 @@ import { Elevation } from "@webiny/ui/Elevation";
 import { ContentModelForm } from "@webiny/app-headless-cms/admin/components/ContentModelForm";
 import { i18n } from "@webiny/app/i18n";
 import { CmsEditorContentTab } from "@webiny/app-headless-cms/types";
+import { useI18N } from "@webiny/app-i18n/hooks/useI18N";
+
 const t = i18n.ns("app-headless-cms/admin/components/editor/tabs/preview");
 
 const formPreviewWrapper = css({
@@ -22,13 +24,14 @@ const style = {
 
 export const PreviewTab: CmsEditorContentTab = ({ activeTab }) => {
     const { data } = useContentModelEditor();
+    const i18n = useI18N();
 
     const { fields } = data;
 
     return (
         <Elevation z={1} className={formPreviewWrapper}>
             {fields && fields.length && activeTab ? (
-                <ContentModelForm contentModel={data} />
+                <ContentModelForm contentModel={data} locale={i18n.getLocale().id} />
             ) : (
                 <div className={style.noFieldsMessage}>
                     {t`Before previewing the form, please add at least one field to the content model.`}