+++ packages/app-headless-cms/src/admin/plugins/fieldRenderers/index.ts
@@ -5,6 +5,19 @@ import richText from "./richText";
 import boolean from "./boolean";
 import dateTime from "./dateTime";
 import file from "./file";
+import radioButtons from "./radioButtons";
+import checkboxes from "./checkboxes";
 import ref from "./ref";
 
-export default [number, text, longText, richText, boolean, dateTime, file, ref];
+export default [
+    number,
+    text,
+    longText,
+    richText,
+    boolean,
+    dateTime,
+    file,
+    ref,
+    radioButtons,
+    checkboxes
+];