+++ packages/app-security/src/hooks/useSecurity.d.ts
@@ -1,0 +1,2 @@
+import { SecurityContextValue } from "../contexts/Security";
+export declare const useSecurity: () => SecurityContextValue;