+++ packages/project-utils/bundling/app/config/paths.js
@@ -65,7 +65,6 @@ const resolveModule = (resolveFn, filePath) => {
 
 // config after eject: we're in ./config/
 module.exports = ({ appIndexJs }) => ({
-    dotenv: resolveApp(".env"),
     appPath: resolveApp("."),
     appBuild: resolveApp("build"),
     appPublic: resolveApp("public"),