+++ packages/project-utils/bundling/app/config/webpackDevServer.config.js
@@ -3,13 +3,12 @@
 const errorOverlayMiddleware = require("react-dev-utils/errorOverlayMiddleware");
 const evalSourceMapMiddleware = require("react-dev-utils/evalSourceMapMiddleware");
 const ignoredFiles = require("react-dev-utils/ignoredFiles");
-const paths = require("./paths");
 const fs = require("fs");
 
 const protocol = process.env.HTTPS === "true" ? "https" : "http";
 const host = process.env.HOST || "0.0.0.0";
 
-module.exports = function(proxy, allowedHost) {
+module.exports = function(proxy, allowedHost, paths) {
     return {
         // WebpackDevServer 2.4.3 introduced a security fix that prevents remote
         // websites from potentially accessing local content through DNS rebinding: