+++ packages/commodo-graphql/CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 4.0.0-beta.15 (2020-06-03)
+
+
+### Bug Fixes
+
+* **project-utils:** remove unused dependencies ([95e06ca](https://github.com/webiny/webiny-js/commit/95e06ca58d88131af665a59041e4355ce1ad16d8))
+
+
+
+
+
+# 4.0.0-beta.14 (2020-06-03)
+
+
+### Bug Fixes
+
+* **ui:** remove unused type import ([9e5a94d](https://github.com/webiny/webiny-js/commit/9e5a94d7b6bb4cb7c44b84c876dd130ec05f6507))
+
+
+
+
+
+# 4.0.0-beta.13 (2020-06-02)
+
+**Note:** Version bump only for package @webiny/commodo-graphql
+
+
+
+
+
+# 4.0.0-beta.12 (2020-06-02)
+
+**Note:** Version bump only for package @webiny/commodo-graphql
+
+
+
+
+
+# 4.0.0-beta.11 (2020-06-02)
+
+
+### Bug Fixes
+
+* 🐛  Add dependencies for packages in app-admin ([0cbd052](https://github.com/webiny/webiny-js/commit/0cbd0526d90bba17f0ef5b00b29a35d84bbd831a))
+
+
+
+
+
+# 4.0.0-beta.10 (2020-06-02)
+
+**Note:** Version bump only for package @webiny/commodo-graphql
+
+
+
+
+
+# 4.0.0-beta.9 (2020-06-02)
+
+**Note:** Version bump only for package @webiny/commodo-graphql
+
+
+
+
+
+# 4.0.0-beta.8 (2020-06-02)
+
+
+### Features
+
+* ✨  Remove unneeded text below password. Add link to checkbox ([39687f4](https://github.com/webiny/webiny-js/commit/39687f42f17c0066c681c16745ab9c37d5759f08))
+
+
+
+
+
 # 4.0.0-beta.7 (2020-06-01)
 
 **Note:** Version bump only for package @webiny/commodo-graphql